Created enum TradeRulesType as a parameter for TradeRules, Deviation or Absolute
Rules can now be passed to TradingRules as absolute figures or deviation. Originally it was only deviation based but it's better to have both options available as inputs and calculate the other one upon initiation.

Also created Enum Signal which contains Buy and Sell, instead of passing strings around as signals