Stores buy and sell operators as enums upon input
instead of calling Operator() when recieving a message, does this beforehand and stores into an enum Operator then uses a switch case when recieving message.
I also refactored the connections class to use arrays instead of lists.

The results significantly brought down the stopwatch from 15-30ms down to 0.3-0.5ms which is a massive improvement and would be acceptable for live trading :)