Added conflict manager
In TradeExecuter I have added two seperate private methods for if it's a single Indicator/TradeRules or if it is multiple and requires conflict management ie. if all the signals agree.

Performs decently for the SingleIndicator method. about 1.5ms
Can probably squeeze a bit more out of that.

Ass for the conflict manager... The basic idea works now but the performance is really terrible. Gonna have to seriously optimize it. But I will do that tomorrow. been up all night